change permissions on ddclient.conf to 600
ddclient complains if you have something other than 600 perms. if prefix == "" then change permissions

+++ untangle-netd/hier/usr/lib/python2.7/netd/ddclient_manager.py
@@ -113,14 +113,16 @@ class DdclientManager:
         self.write_ddclient_config_file( settings, prefix, verbosity )
         self.write_ddclient_default_file( settings, prefix, verbosity )
 
-        if settings.get('dynamicDnsServiceEnabled'):
-            os.system('/usr/sbin/update-rc.d ddclient defaults >/dev/null 2>&1')
-            os.system('/etc/init.d/ddclient restart >/dev/null 2>&1')
-        else:
-            os.system('/usr/sbin/update-rc.d -f ddclient remove >/dev/null 2>&1')
-            # this doesn't work because it checks /etc/default/ddclient first
-            # use killall instead
-            # os.system('/etc/init.d/ddclient stop >/dev/null 2>&1')
-            os.system('killall ddclient >/dev/null 2>&1')
+        if prefix == "":
+            if settings.get('dynamicDnsServiceEnabled'):
+                os.system('chmod 600 /etc/ddclient.conf')
+                os.system('/usr/sbin/update-rc.d ddclient defaults >/dev/null 2>&1')
+                os.system('/etc/init.d/ddclient restart >/dev/null 2>&1')
+            else:
+                os.system('/usr/sbin/update-rc.d -f ddclient remove >/dev/null 2>&1')
+                # this doesn't work because it checks /etc/default/ddclient first
+                # use killall instead
+                # os.system('/etc/init.d/ddclient stop >/dev/null 2>&1')
+                os.system('killall ddclient >/dev/null 2>&1')
 
         return
